Use a DOMString for encryption scheme (#12)
Using a DOMString instead of an enum for the encryptionScheme member
allows for safer extension of encryption schemes. Without using a
DOMString, invalid enum values will result in a TypeError.

An example of why this is problematic: if a new encryption scheme, fooo,
is added, and an application checks for support using the API, an older
User Agent may throw a TypeError as fooo is not a valid enum value.
Using a DOMString avoids this, and User Agents can treat configurations
with unrecognized schemes as unsupported. This allows the call to
requestMediaKeySystemAccess() to succeed, provided other configurations
with supported schemes are provided.

Closes #8